+++ modules/services/preferences/src/main/java/au/com/shiftyjelly/pocketcasts/preferences/Settings.kt
@@ -400,7 +400,6 @@ interface Settings {
     val autoDownloadUpNext: ReadWriteSetting<Boolean>
     val autoDownloadOnFollowPodcast: UserSetting<Boolean>
     val autoDownloadNewEpisodes: ReadWriteSetting<Int>
-    val pendingAutoDownloadEpisodeUuids: ReadWriteSetting<List<String>>
 
     val artworkConfiguration: UserSetting<ArtworkConfiguration>